make: avoid generating objects on error
Even in pipefail mode, the bash pipeline is executed up to its end:
https://stackoverflow.com/questions/73162200/bash-set-eo-pipefail-not-immediately-exiting
Also cpp is processing its input and generates an output until it
hits an error.
This combined, may lead to an incomplete but syntactically valid
input be presented to gcc, even if error was encountered.
And bad object file is then generated.

Use intermediate file as a solution.