API v17: infra for pluggable ELF parsers [#23]
This allows to stop adding crap (like new ELF loader per every
new ELF format) and off-load that task into dosemu2 if he needs
the direct ELF loading capability.
Note that this parser can't help with loading an ELF intended
for static linking, as the parser itself is invoked by the dynamic
payload (which, in case of a static ELF, is embedded into that
ELF, so can't call the parser to parse itself).