glue go64 loader with new ELF infra [closes #16]
Now it is finally possible to load most of the supported ELF
formats (modulo those with compactified layout and with static linking)
via go64 API.

This is a huge overhaul of the entire ELF infrastructure, but
at the same time its a simplest possible approach. If someone
wants to expose the remaining ELF formats to go64, then that
would be 100 times more work, so not worth the trouble.