From: "J.W. Jagersma" <jwjagersma@gmail.com> Subject: PATCH: djasm: fix sortsyms() prototype for gcc 15
With gcc 15 now compiling in C23 mode by default, it no longer allows
incomplete function prototypes (an empty parameter list is now the same
as (void), like in C++).  This is used for one function in djasm, which
now fails to compile.  The small patch below is all that's needed to fix
it.  For the other hostbin utilities, no changes are necessary.